laba5 final refactoring

+++ src/main.rs
@@ -23,26 +23,25 @@ const _MASK: [[i32; 3]; 3] =
 
 
 fn _color_calc(x: u32, y: u32, index: usize, img: &DynamicImage) -> u8 {
-    let mut result: i64
-        = ( img.get_pixel(x, y)[index] as i32 * _MASK[1][1] as i32
-        + img.get_pixel(x + 1, y)[index] as i32 * _MASK[2][1] as i32
-        + img.get_pixel(x, y + 1)[index] as i32 * _MASK[1][2] as i32
-        + img.get_pixel(x - 1, y)[index] as i32 * _MASK[0][1] as i32
-        + img.get_pixel(x, y - 1)[index] as i32 * _MASK[1][0] as i32
-        + img.get_pixel(x - 1, y - 1)[index] as i32 * _MASK[0][0] as i32
-        + img.get_pixel(x + 1, y + 1)[index] as i32 * _MASK[2][2] as i32
-        + img.get_pixel(x - 1, y + 1)[index] as i32 * _MASK[0][2] as i32
-        + img.get_pixel(x + 1, y - 1)[index] as i32 * _MASK[2][0] as i32 ) as i64;
+    let mut result: i32 = img.get_pixel(x, y)[index] as i32 * _MASK[1][1]
+        + img.get_pixel(x + 1, y)[index] as i32 * _MASK[2][1] 
+        + img.get_pixel(x, y + 1)[index] as i32 * _MASK[1][2] 
+        + img.get_pixel(x - 1, y)[index] as i32 * _MASK[0][1]
+        + img.get_pixel(x, y - 1)[index] as i32 * _MASK[1][0]
+        + img.get_pixel(x - 1, y - 1)[index] as i32 * _MASK[0][0] 
+        + img.get_pixel(x + 1, y + 1)[index] as i32 * _MASK[2][2]
+        + img.get_pixel(x - 1, y + 1)[index] as i32 * _MASK[0][2] 
+        + img.get_pixel(x + 1, y - 1)[index] as i32 * _MASK[2][0];
 
-    let wight: i128 = (_MASK[1][1] as i128 * _MASK[2][1] as i128 * _MASK[1][2] as i128 * _MASK[0][1] as i128 * _MASK[1][0] as i128 * _MASK[0][0] as i128 * _MASK[2][2] as i128 * _MASK[0][2] as i128 * _MASK[2][0] as i128 + 1);
+    let wight: i128 = _MASK[1][1] as i128 * _MASK[2][1] as i128 * _MASK[1][2] as i128 * _MASK[0][1] as i128 * _MASK[1][0] as i128 * _MASK[0][0] as i128 * _MASK[2][2] as i128 * _MASK[0][2] as i128 * _MASK[2][0] as i128 + 1;
 
     if result > 255 {
-        result = (255 / wight) as i64;
+        result = (255 / wight) as i32;
     } else {
         if result < 0 {
             result = result * -1;
             if result > 255 {
-                result = (255 / wight) as i64;
+                result = (255 / wight) as i32;
             }
         }
     }
